# docs/swagger.yaml
swagger: "2.0"
info:
  description: "API para acesso aos dados da Secretaria Municipal de Saúde sobre COVID-19"
  version: "1.0.0"
  title: "API Covid19 Vilhena - RO"
  termsOfService: "https://api.covid19.adsvilhena.ninja/terms/"
  contact:
    email: "marco.andrade@ifro.edu.br"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "api.covid19.adsvilhena.ninja"
basePath: "/v1"
schemes:
- "https"
paths:
  /bairros:
    get:
      tags:
      - "Bairro"
      description: "Lista os  bairros"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Bairro"
        "404":
          description: "Nenhum dado encontrado"

    post:
      tags:
      - "Bairro"
      description: "Inseri os dados de um novo bairro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Bairro"
        required: true
        schema:
          $ref: "#/definitions/Bairro"
      responses:
        "200":
          description: "Novo bairro inserido com sucesso"
          schema:
            $ref: "#/definitions/Bairro"
      security:
        - api_key: []

    put:
      tags:
      - "Bairro"
      description: "Atualiza os dados de um bairro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Bairro"
        required: true
        schema:
          $ref: "#/definitions/Bairro"
      responses:
        "200":
          description: "Bairro alterado com sucesso"
          schema:
            $ref: "#/definitions/Bairro"
        "404":
          description: "Nenhum bairro encontrado para este id"
      security:
        - api_key: []
    
    delete:
      tags:
      - "Bairro"
      description: "Deleta um bairro"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Bairro"
        required: true
        schema:
          $ref: "#/definitions/Bairro"
      responses:
        "200":
          description: "Bairro deletado com sucesso"
        "404":
          description: "Nenhum bairro encontrado para este id"
      security:
        - api_key: []
  /casos:
    get:
      tags:
      - "Casos"
      description: "Lista os dados dos casos"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Caso"
        "404":
          description: "Nenhum dado encontrado"

    post:
      tags:
      - "Casos"
      description: "Inseri os dados dos casos"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Caso"
        required: true
        schema:
          $ref: "#/definitions/Caso"
      responses:
        "200":
          description: "Sexos inseridos com sucesso"
          schema:
            $ref: "#/definitions/Caso"
      security:
        - api_key: []

    delete:
      tags:
      - "Casos"
      description: "Deleta um dado sobre caso"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto Caso"
        required: true
        schema:
          $ref: "#/definitions/Caso"
      responses:
        "200":
          description: "Dado sobre casos deletado com sucesso"
        "404":
          description: "Nenhum dado sobre casos encontrado para este id"
      security:
        - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Bairro:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      nome:
        type: "string"
      casos_ativos:
        type: "integer"
        format: "int64"
      coordenadas:
        type: "integer"
        # type: "array"
      data:
        type: "string"
  Caso:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      dia:
        type: "integer"
        format: "int64"
      data:
        type: "string"
        # format: "date"
      ativos:
        type: "integer"
        format: "int64"
      suspeitos:
        type: "integer"
        format: "int64"
      confirmados:
        type: "integer"
        format: "int64"
      recuperados:
        type: "integer"
        format: "int64"
      descartados:
        type: "integer"
        format: "int64"
      internados:
        type: "integer"
        format: "int64"
      novos:
        type: "integer"
        format: "int64"
      obitos:
        type: "integer"
        format: "int64"
...